feat: ensure DeleteQuestionUseCase does not delete question from another author

+++ src/domain/forum/application/usecases/delete-question.ts
@@ -1,6 +1,7 @@
 import { QuestionsRepository } from '../repositories/questions-repository'
 
 interface DeleteQuestionUseCaseRequest {
+  authorId: string
   questionId: string
 }
 
@@ -10,12 +11,17 @@ export class DeleteQuestionUseCase {
   constructor(private questionRepository: QuestionsRepository) {}
 
   async execute({
+    authorId,
     questionId,
   }: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
     const question = await this.questionRepository.findById(questionId)
 
     if (!question) {
       throw new Error('Question not found.')
+    }
+
+    if (authorId !== question.authorId.toString()) {
+      throw new Error('Now allowed.')
     }
 
     await this.questionRepository.delete(question)
